pyproject.toml
[project]
name = "nlp-reco"
version = "0.1.0"
description = "NLP-powered hotel recommendation system using BERTopic and Turkish morphological analysis"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pandas",
    "numpy",
    "matplotlib",
    "seaborn",
    "wordcloud",
    "spacy",
    "bertopic",
    "langdetect",
    "jpype1>=1.4.0,<1.6.0",  # 1.6.0 breaks with Java 8; 1.5.x is stable
    "scikit-learn",
    "tqdm",
]

[project.optional-dependencies]
notebook = [
    "jupyter",
    "ipykernel",
]
dev = [
    "pytest",
]

[tool.uv]
dev-dependencies = [
    "pytest",
    "jupyter",
    "ipykernel",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
# This is a scripts-based project; include both source packages
packages = ["src", "utils"]
